# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CollegeCompass API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SignupInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface College {
  id: number;
  name: string;
  city: string;
  state: string;
  /** IIT, NIT, IIIT, Private, Government */
  type: string;
  /** Annual fees in INR */
  fees: number;
  rating: number;
  /** @nullable */
  logoUrl: string | null;
  shortDescription: string;
  /** @nullable */
  establishedYear?: number | null;
  /** @nullable */
  website?: string | null;
  createdAt?: string;
}

export interface Course {
  id: number;
  collegeId: number;
  name: string;
  /** e.g. "4 years" */
  duration: string;
  /** Annual fees in INR */
  fees: number;
  /** @nullable */
  description?: string | null;
}

export interface Review {
  id: number;
  collegeId: number;
  authorName: string;
  rating: number;
  title: string;
  content: string;
  createdAt: string;
}

export interface CollegeDetail {
  id: number;
  name: string;
  city: string;
  state: string;
  type: string;
  fees: number;
  rating: number;
  /** @nullable */
  logoUrl: string | null;
  shortDescription: string;
  /** @nullable */
  about: string | null;
  /** @nullable */
  establishedYear?: number | null;
  /** @nullable */
  website?: string | null;
  /**
     * Average placement package in LPA
     * @nullable
     */
  avgPackage: number | null;
  /**
     * Highest package in LPA
     * @nullable
     */
  highestPackage: number | null;
  recruiters: string[];
  courses: Course[];
  reviews: Review[];
}

export interface CollegeListResponse {
  colleges: College[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface ReviewInput {
  /** @minLength 2 */
  authorName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  content: string;
}

export interface SavedComparison {
  id: number;
  userId: number;
  collegeIds: number[];
  label: string;
  createdAt: string;
}

export interface SavedComparisonInput {
  /**
     * @minItems 2
     * @maxItems 3
     */
  collegeIds: number[];
  /** @minLength 1 */
  label: string;
}

export type PredictorInputExamType = typeof PredictorInputExamType[keyof typeof PredictorInputExamType];


export const PredictorInputExamType = {
  JEE_MAIN: 'JEE_MAIN',
  JEE_ADVANCED: 'JEE_ADVANCED',
  CUET: 'CUET',
  NEET: 'NEET',
  CAT: 'CAT',
  GATE: 'GATE',
} as const;

export interface PredictorInput {
  examType: PredictorInputExamType;
  /** @minimum 1 */
  rank: number;
}

export type PredictorResultChance = typeof PredictorResultChance[keyof typeof PredictorResultChance];


export const PredictorResultChance = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface PredictorResult {
  college: College;
  chance: PredictorResultChance;
  /** Score from 0 to 100 */
  matchScore: number;
}

export interface SaveCollegeInput {
  collegeId: number;
}

export interface SavedCollegeRecord {
  id: number;
  userId: number;
  collegeId: number;
  createdAt: string;
}

export interface StatsSummary {
  totalColleges: number;
  totalCourses: number;
  totalReviews: number;
  avgRating: number;
}

export interface CityCount {
  city: string;
  state: string;
  count: number;
}

export type ListCollegesParams = {
search?: string;
city?: string;
type?: string;
minFees?: number;
maxFees?: number;
minRating?: number;
sortBy?: ListCollegesSortBy;
sortOrder?: ListCollegesSortOrder;
page?: number;
limit?: number;
};

export type ListCollegesSortBy = typeof ListCollegesSortBy[keyof typeof ListCollegesSortBy];


export const ListCollegesSortBy = {
  name: 'name',
  rating: 'rating',
  fees: 'fees',
  createdAt: 'createdAt',
} as const;

export type ListCollegesSortOrder = typeof ListCollegesSortOrder[keyof typeof ListCollegesSortOrder];


export const ListCollegesSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type GetComparisonParams = {
/**
 * Comma-separated college IDs (2-3)
 */
ids: string;
};

export type GetTopCollegesParams = {
limit?: number;
};