# lib/api-zod/src/generated/types/savedCollegeRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CollegeCompass API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SavedCollegeRecord {
  id: number;
  userId: number;
  collegeId: number;
  createdAt: Date;
}